WIP review feedback

+++ server/src/main/java/io/deephaven/server/table/stats/FloatChunkedNumericalStats.java
@@ -76,7 +76,7 @@ public class FloatChunkedNumericalStats implements ChunkedNumericalStatsKernel {
 
                     chunkedSum += val;
                     chunkedAbsSum += absVal;
-                    chunkedSqrdSum += val * val;
+                    chunkedSqrdSum += (double) val * (double) val;
                 }
 
                 sum += chunkedSum;
